+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Adds support for @query type (user defined functions)
 - Adds support for recursive references
 - Removes `get`, `post`, `put`, `patch` and `delete` methods from `Client` class
+- Adds `abort` function
 
 ## 1.1.2 (July 06, 2017)
 - Fix typescript declaration file to include default types